Avoid 64bit division in multilist index functions.
Number of sublists in multilists equal to number of CPUs is typically
small even or even power of 2.  It makes almost no sense to use full
slow 64bit division.

Sponsored-By: iXsystems, Inc.

+++ module/zfs/dmu_objset.c
@@ -399,7 +399,15 @@ static unsigned int
 dnode_multilist_index_func(multilist_t *ml, void *obj)
 {
 	dnode_t *dn = obj;
-	return (dnode_hash(dn->dn_objset, dn->dn_object) %
+
+	/*
+	 * The low order bits of the hash value are thought to be
+	 * distributed evenly. Otherwise, in the case that the multilist
+	 * has a power of two number of sublists, each sublists' usage
+	 * would not be evenly distributed. In this context full 64bit
+	 * division would be a waste of time, so limit it to 32 bits.
+	 */
+	return ((unsigned int)dnode_hash(dn->dn_objset, dn->dn_object) %
 	    multilist_get_num_sublists(ml));
 }
